Attempt to undo moving files out of /dataset for panel since not yet using
