+++ src/components/dataset/query/sidebar/panels/InfoViewDatasetAccess.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { withStyles } from '@material-ui/core/styles';
 import { Grid, Link, Typography } from '@material-ui/core';
 import { connect } from 'react-redux';
-import DatasetAccess from '../../../../dataset/DatasetAccess';
+import DatasetAccess from '../../../DatasetAccess';
 
 const styles = (theme) => ({
   root: {
