+++ src/components/dataset/query/sidebar/panels/ShareSnapshot.jsx
